Make ability modifier a point-free composed function
While this looks weird if you're not used to it, you'll see this style
a lot in haskell. The confusion of `f = flip div 2` (equivalent to `f x
= x \`div\` 2`) is considered acceptable when it comes with the
expressiveness of "abilityModifier is the composition of minusTen and
halfOf". Naturally, this is a style decision, but the overhead to
learning to read functions like this is relatively low when compared to
the number of time you might read function bodies full of lengthy symbol
names.