Dubious refactors to point-free style
Here are a few less-desirable refactors to point-free style. While
definitions such as "damageForAttack = max of 1 and rawDamageForAttack"
seems like a really great function definition that accepts a character
in a point-free way, due to calling order in client code, we have a
calling order that is unclear due to the necessary compositions. This is
probably something we want to avoid if we're ever worried about making
changes to this definition, since it will force someone to have to read
and think about it.

The refactor of currentLevel is much less confusing, since the
definitions are rather declarative, even if they do have an oddly-placed
composition.

- currentLevel is 1 + the levelsOfExperience
- levelsofexperience is the expPerLevel with currentExperience
- expPerLevel is a private function that divides by the amount of
  experience needed to get to another level

We'll undo damageForAttack and put it in a nicer form next.